fix: cleanup of how child state is cleared
